frontend/ rename BoardModal to AppBoardModal
